Add previous behaviour to docker container
The docker container continues to work as before if no CRON variable is set.
If execution parameters ere defined through the OPTIONS variable, the container will use those in conujunction with the directories mounted to /mnt/input and /mnt/output. Both of those mount point are declared in the dockerfile.

+++ entrypoint.sh
@@ -1,11 +1,28 @@
 #!/usr/bin/env bash
 
-if [ -f /tmp/phockup.lockfile ]; then
-  rm /tmp/phockup.lockfile
+# If the CRON variable is empty, phockup gets executed once as command line tool
+if [ -z "$CRON" ]; then
+  if [ -z "$OPTIONS" ]; then
+    phockup $@
+  else
+    phockup /mnt/input /mnt/output $OPTIONS
+  fi
+
+# When CRON is not empty, phockup will run in a cron job until the container is stopped.
+else
+  if [ -f /tmp/phockup.lockfile ]; then
+    rm /tmp/phockup.lockfile
+  fi
+  
+  if [ -z "$OPTIONS" ]; then
+    CRON_COMMAND="$CRON flock -n /tmp/phockup.lockfile phockup $@"
+  else
+    CRON_COMMAND="$CRON flock -n /tmp/phockup.lockfile phockup /mnt/input /mnt/output $OPTIONS"
+  fi  
+
+  echo "$CRON_COMMAND" >> /etc/crontabs/root
+  echo "cron job has been set up with command: $CRON_COMMAND"
+
+  crond -f -d 8
 fi
 
-CRON_COMMAND="$CRON flock -n /tmp/phockup.lockfile phockup /mnt/input /mnt/output $OPTIONS"
-echo "$CRON_COMMAND" >> /etc/crontabs/root
-echo "cron job has been set up with command: $CRON_COMMAND"
-
-crond -f -d 8
